Deploy RC 228 to Prod

+++ spec/services/irs_attempts_api/attempt_event_spec.rb
@@ -54,7 +54,7 @@ RSpec.describe IrsAttemptsApi::AttemptEvent do
         'subject_type' => 'session', 'session_id' => 'test-session-id',
       )
       expect(event_data['foo']).to eq('bar')
-      expect(event_data['occurred_at']).to eq(occurred_at.to_i)
+      expect(event_data['occurred_at']).to eq(occurred_at.to_f)
     end
   end
